requirements.txt
fastapi
pymongo
python-dotenv

